feat:adds edit profile modal and does basic restructure on app.jsx
